initialise portfolio module
